chore: move presentation related code together
