+++ src/presenter.rs
@@ -11,13 +11,16 @@ use crate::{
     code::execute::SnippetExecutor,
     commands::listener::{Command, CommandListener},
     custom::KeyBindingsConfig,
-    diff::PresentationDiffer,
     export::ImageReplacer,
     markdown::parse::{MarkdownParser, ParseError},
-    presentation::{Presentation, RenderAsyncState},
-    processing::builder::{BuildError, PresentationBuilder, PresentationBuilderOptions, Themes},
+    presentation::{
+        Presentation,
+        builder::{BuildError, PresentationBuilder, PresentationBuilderOptions, Themes},
+        diff::PresentationDiffer,
+    },
     render::{
         draw::{ErrorSource, RenderError, RenderResult, TerminalDrawer},
+        operation::RenderAsyncState,
         properties::WindowSize,
         validate::OverflowValidator,
     },
