--- src/processing/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod builder;
