+++ src/render/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod draw;
 pub(crate) mod engine;
 pub(crate) mod layout;
+pub(crate) mod operation;
 pub(crate) mod properties;
 pub(crate) mod text;
 pub(crate) mod validate;
